+++ networking/eth2/src/test/java/tech/pegasys/teku/networking/eth2/gossip/topics/Eth2GossipTopicFilterTest.java
@@ -245,7 +245,7 @@ class Eth2GossipTopicFilterTest {
   @TestTemplate
   void shouldNotConsiderExecutionProofSubnetsRelevantByDefault() {
     assumeThat(nextSpecMilestone).isEqualTo(ELECTRA);
-    for (int i = 0; i < Constants.MAX_EXECUTION_PROOF_SUBNETS.intValue(); i++) {
+    for (int i = 0; i < Constants.MAX_EXECUTION_PROOF_SUBNETS; i++) {
       assertThat(filter.isRelevantTopic(getTopicName(getExecutionProofSubnetTopicName(i))))
           .isFalse();
     }
@@ -257,7 +257,7 @@ class Eth2GossipTopicFilterTest {
         P2PConfig.builder().specProvider(spec).executionProofTopicEnabled(true).build();
     filter = new Eth2GossipTopicFilter(recentChainData, SSZ_SNAPPY, spec, p2pConfigOverwritten);
     assumeThat(nextSpecMilestone).isEqualTo(ELECTRA);
-    for (int i = 0; i < Constants.MAX_EXECUTION_PROOF_SUBNETS.intValue(); i++) {
+    for (int i = 0; i < Constants.MAX_EXECUTION_PROOF_SUBNETS; i++) {
       assertThat(filter.isRelevantTopic(getTopicName(getExecutionProofSubnetTopicName(i))))
           .isTrue();
     }
